refactor(remix): separate box adapter from effects renderer
