+++ packages/remix/lib/src/rendering/remix_box_adapter.dart
@@ -1,10 +1,10 @@
 part of 'remix_box_effects.dart';
 
-/// Routes a resolved Mix [Box] surface through the simplest valid renderer.
+/// Routes a resolved Mix [Box] through the simplest compatible renderer.
 ///
-/// Ordinary surfaces remain real Mix Boxes. Only non-empty advanced effects
-/// use [_RemixAdvancedSurface], while CSS-style negative margins are handled by
-/// the independent outer-layout adapter.
+/// Ordinary boxes stay on Mix's standard renderer. Advanced effects use the
+/// dedicated [_RemixBoxEffectsRenderer], while CSS-style negative margins are
+/// handled by the independent outer-layout adapter.
 @internal
 final class RemixBoxAdapter extends StatelessWidget {
   const RemixBoxAdapter({
@@ -26,7 +26,7 @@ final class RemixBoxAdapter extends StatelessWidget {
       styleSpec: styleSpec,
       builder: (context, spec) {
         final margin = spec.margin;
-        if (effects.isEmpty && (margin == null || margin.isNonNegative)) {
+        if (_canUseStandardMixRendering(effects: effects, margin: margin)) {
           return Box(
             styleSpec: StyleSpec(spec: spec),
             child: child,
@@ -42,9 +42,9 @@ final class RemixBoxAdapter extends StatelessWidget {
           );
         } else {
           _validateBoxDecoration(spec.decoration);
-          inner = _RemixAdvancedSurface(
+          inner = _RemixBoxEffectsRenderer(
             spec: innerSpec,
-            containerEffects: effects,
+            effects: effects,
             child: child,
           );
         }
@@ -83,8 +83,10 @@ final class RemixFlexBoxAdapter extends StatelessWidget {
       styleSpec: styleSpec,
       builder: (context, spec) {
         final box = spec.box?.spec;
-        final margin = box?.margin;
-        if (effects.isEmpty && (margin == null || margin.isNonNegative)) {
+        if (_canUseStandardMixRendering(
+          effects: effects,
+          margin: box?.margin,
+        )) {
           final resolved = StyleSpec(spec: spec);
           return switch (direction) {
             Axis.horizontal => RowBox(styleSpec: resolved, children: children),
@@ -123,47 +125,54 @@ final class RemixFlexBoxAdapter extends StatelessWidget {
   }
 }
 
-void _validateEffects(RemixBoxEffectsSpec containerEffects) {
-  if (!containerEffects.backdropBlur.isFinite) {
+bool _canUseStandardMixRendering({
+  required RemixBoxEffectsSpec effects,
+  EdgeInsetsGeometry? margin,
+}) => effects.isEmpty && (margin == null || margin.isNonNegative);
+
+/// Validates at the adapter boundary so every rendering path rejects the same
+/// unsupported values before choosing an implementation.
+void _validateEffects(RemixBoxEffectsSpec effects) {
+  if (!effects.backdropBlur.isFinite) {
     throw FlutterError(
       'RemixBoxEffectsSpec.backdropBlur must be finite. '
-      'Received ${containerEffects.backdropBlur}.',
+      'Received ${effects.backdropBlur}.',
     );
   }
-  if (containerEffects.backdropBlur < 0) {
+  if (effects.backdropBlur < 0) {
     throw FlutterError(
       'RemixBoxEffectsSpec.backdropBlur must be non-negative. '
-      'Received ${containerEffects.backdropBlur}.',
+      'Received ${effects.backdropBlur}.',
     );
   }
-  if (!containerEffects.outline.width.isFinite) {
+  if (!effects.outline.width.isFinite) {
     throw FlutterError(
       'RemixBoxEffectsSpec.outline.width must be finite. '
-      'Received ${containerEffects.outline.width}.',
+      'Received ${effects.outline.width}.',
     );
   }
-  if (containerEffects.outline.width < 0) {
+  if (effects.outline.width < 0) {
     throw FlutterError(
       'RemixBoxEffectsSpec.outline.width must be non-negative. '
-      'Received ${containerEffects.outline.width}.',
+      'Received ${effects.outline.width}.',
     );
   }
-  if (!containerEffects.outlineOffset.isFinite) {
+  if (!effects.outlineOffset.isFinite) {
     throw FlutterError(
       'RemixBoxEffectsSpec.outlineOffset must be finite. '
-      'Received ${containerEffects.outlineOffset}.',
+      'Received ${effects.outlineOffset}.',
     );
   }
-  if (containerEffects.outline.style != BorderStyle.none &&
-      containerEffects.outline.strokeAlign != BorderSide.strokeAlignInside) {
+  if (effects.outline.style != BorderStyle.none &&
+      effects.outline.strokeAlign != BorderSide.strokeAlignInside) {
     throw FlutterError(
       'Remix box effects outlines must use BorderSide.strokeAlignInside.',
     );
   }
-  if (containerEffects.behindContent case final layer?) {
+  if (effects.behindContent case final layer?) {
     _validateEffectLayer(layer, 'behindContent');
   }
-  if (containerEffects.overContent case final layer?) {
+  if (effects.overContent case final layer?) {
     _validateEffectLayer(layer, 'overContent');
   }
 }
@@ -223,146 +232,4 @@ void _validateBoxDecoration(Decoration? decoration) {
     'Remix box effects require a rectangular BoxDecoration. '
     'ShapeDecoration and BoxShape.circle are not supported.',
   );
-}
-
-/// Internal Box equivalent that inserts box effect layers at the decoration
-/// boundary, inside constraints and outside CSS margin and transforms.
-class _RemixAdvancedSurface extends StatelessWidget {
-  const _RemixAdvancedSurface({
-    required this.spec,
-    required this.containerEffects,
-    this.child,
-  });
-
-  final BoxSpec spec;
-  final RemixBoxEffectsSpec containerEffects;
-  final Widget? child;
-
-  @override
-  Widget build(BuildContext context) {
-    Widget? current = child;
-    if (child == null) {
-      current = spec.constraints == null || !spec.constraints!.isTight
-          ? LimitedBox(
-              maxWidth: 0,
-              maxHeight: 0,
-              child: ConstrainedBox(constraints: const BoxConstraints.expand()),
-            )
-          : const SizedBox.shrink();
-    } else if (spec.alignment != null) {
-      current = Align(alignment: spec.alignment!, child: current);
-    }
-
-    final decorationPadding = spec.decoration?.padding;
-    final padding = switch ((spec.padding, decorationPadding)) {
-      (null, final value) => value,
-      (final value, null) => value,
-      (final value?, final decoration?) => value.add(decoration),
-    };
-    if (padding != null) current = Padding(padding: padding, child: current);
-
-    final decoration =
-        spec.decoration as BoxDecoration? ?? const BoxDecoration();
-    final textDirection = Directionality.maybeOf(context);
-    final borderRadius = (decoration.borderRadius ?? BorderRadius.zero).resolve(
-      textDirection,
-    );
-    final behindContent =
-        containerEffects.behindContent ?? const RemixBoxEffectLayerSpec();
-
-    // Box fill -> advanced behindContent -> border -> child.
-    if (decoration.border case final border?) {
-      current = CustomPaint(
-        painter: _RemixBoxBorderPainter(
-          border: border,
-          borderRadius: borderRadius,
-          textDirection: textDirection,
-        ),
-        child: current,
-      );
-    }
-    if (!behindContent.isEmpty) {
-      current = CustomPaint(
-        painter: _RemixBoxEffectLayerPainter(
-          spec: behindContent,
-          borderRadius: borderRadius,
-          textDirection: textDirection,
-          phase: _BoxEffectPaintPhase.inner,
-        ),
-        child: current,
-      );
-    }
-    current = DecoratedBox(
-      decoration: _backgroundDecoration(decoration),
-      child: current,
-    );
-
-    if (containerEffects.backdropBlur > 0) {
-      current = ClipRRect(
-        borderRadius: borderRadius,
-        clipBehavior: Clip.antiAlias,
-        child: BackdropFilter(
-          filter: ui.ImageFilter.blur(
-            sigmaX: containerEffects.backdropBlur,
-            sigmaY: containerEffects.backdropBlur,
-          ),
-          child: current,
-        ),
-      );
-    }
-
-    // Clip decorated content without clipping box effect shadows or outlines.
-    if (spec.clipBehavior case final clip? when clip != Clip.none) {
-      current = ClipPath(
-        clipper: _RemixDecorationClipper(
-          textDirection: textDirection,
-          decoration: decoration,
-        ),
-        clipBehavior: clip,
-        child: current,
-      );
-    }
-
-    if (_decorationShadow(decoration) case final shadow?) {
-      current = DecoratedBox(decoration: shadow, child: current);
-    }
-    if (!behindContent.isEmpty) {
-      current = CustomPaint(
-        painter: _RemixBoxEffectLayerPainter(
-          spec: behindContent,
-          borderRadius: borderRadius,
-          textDirection: textDirection,
-          phase: _BoxEffectPaintPhase.outer,
-        ),
-        child: current,
-      );
-    }
-    if (spec.foregroundDecoration case final overContent?) {
-      current = DecoratedBox(
-        decoration: overContent,
-        position: DecorationPosition.foreground,
-        child: current,
-      );
-    }
-    final hasForeground =
-        containerEffects.overContent != null &&
-        !containerEffects.overContent!.isEmpty;
-    final hasOutline =
-        containerEffects.outline.style != BorderStyle.none &&
-        containerEffects.outline.width > 0;
-    if (hasForeground || hasOutline) {
-      current = CustomPaint(
-        foregroundPainter: _RemixBoxEffectsForegroundPainter(
-          containerEffects: containerEffects,
-          borderRadius: borderRadius,
-          textDirection: textDirection,
-        ),
-        child: current,
-      );
-    }
-    if (spec.constraints case final constraints?) {
-      current = ConstrainedBox(constraints: constraints, child: current);
-    }
-    return current;
-  }
 }
