+++ packages/remix/lib/src/rendering/remix_box_effects.dart
@@ -9,6 +9,7 @@ import 'package:mix/mix.dart';
 import 'remix_blend_mode.dart';
 
 part 'remix_box_effects_model.dart';
-part 'remix_box_effects_adapter.dart';
+part 'remix_box_adapter.dart';
 part 'remix_box_outer_layout.dart';
+part 'remix_box_effects_renderer.dart';
 part 'remix_box_effects_painters.dart';
